refactor(http): extract server lifecycle to SocketHTTPServer-core.c
Extract ~700 LOC of server lifecycle and event loop functions:

- Server lifecycle: config_defaults, new, free, start, stop
- Handler registration: set_handler, set_rate_limit, set_validator,
  add_middleware, set_error_handler
- Event loop: server_accept_clients, server_cleanup_timed_out, fd, poll, process
- Drain/shutdown: drain, drain_poll, drain_wait, drain_remaining_ms,
  set_drain_callback, state

Add forward declarations to SocketHTTPServer-private.h for cross-module calls.
Change server_check_connection_timeout and server_process_client_event
from static to non-static for visibility.

Fixes #2627
