+++ CMakeLists.txt
@@ -377,6 +377,7 @@ set(LIB_SOURCES
     src/http/SocketHTTPClient-async.c
     src/http/SocketHTTPServer.c
     src/http/SocketHTTPServer-connections.c
+    src/http/SocketHTTPServer-core.c
     src/http/SocketHTTPServer-static.c
     src/http/SocketHTTPServer-h2.c
     src/http/SocketHTTPServer-metrics.c
